switch to correct getParenthesizedText
